+++ test/createJavascript.spec.ts
@@ -69,6 +69,57 @@ describe("Javascript project generator", () => {
     });
   });
 
+  describe("Generate a basic project in the same directory", () => {
+    before(async () => {
+      simple.mock(inquirer, "prompt").callFn(() => {
+        return Promise.resolve({
+          appName: "my skill",
+          author: "Rain",
+          newDir: false,
+          language: "javascript",
+          voxaCli: false,
+          canFulfill: true,
+          analytics: [],
+          platform: ["all"]
+        });
+      });
+      await action();
+    });
+
+    it("should have a src folder", async () => {
+      const filePath = getFilePath("src");
+      const pathExists = await fs.pathExists(filePath);
+      expect(pathExists).to.be.true;
+    });
+
+    it("should have a server.js file", async () => {
+      const filePath = getFilePath("server.js");
+      const pathExists = await fs.pathExists(filePath);
+      expect(pathExists).to.be.true;
+    });
+
+    it("should have a README file with the name of the skill", async () => {
+      const filePath = getFilePath("README.md");
+      const pathExists = await fs.pathExists(filePath);
+      const fileContent = await fs.readFile(filePath, "utf8");
+      expect(pathExists).to.be.true;
+      expect(fileContent).to.contain("my skill");
+    });
+
+    it("should have a package.json file with the name of the skill and author", async () => {
+      const filePath = getFilePath("package.json");
+      const fileContent = await fs.readFile(filePath, "utf8");
+      expect(fileContent).to.contain('"name": "my-skill",');
+      expect(fileContent).to.contain('"author": "Rain",');
+    });
+
+    it("should have a test folder", async () => {
+      const filePath = getFilePath("test");
+      const pathExists = await fs.pathExists(filePath);
+      expect(pathExists).to.be.true;
+    });
+  });
+
   describe("Generate a Javascript project with voxa-cli", () => {
     before(async () => {
       simple.mock(inquirer, "prompt").callFn(() => {