+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added prompt to check if the project should be generated in the same folder or not.
+
 ## [2.2.0] - 2019-06-10
 
 ### Added